Created compound manifest file to automatically run aide file integrity tool
As i get down to the remaining STIGs, I left a lot of site specific STIGS open. I am going to be more flexible with my rules, and  create variables that need to be populated for some STIGS to work. For example, some STIGS need an email address to send results to, so that will be a new variable in the params.pp file